requirements.txt
discord.py==1.7.1
aiosqlite==0.17.0
